refactor compare page to server component and validate terms
